fix bad malloc

+++ torch_xla/csrc/init_python_bindings.cpp
@@ -254,7 +254,7 @@ at::Tensor DynamicView(const at::Tensor& input,
   return bridge::AtenFromXlaTensor(std::move(result));
 }
 
-at::Tensor ReinterpretCast4bit(const at::Tensor& input, const at::Tensor& weight, const std::vector<int8_t>& int4_weight_values) {
+at::Tensor ReinterpretCast4bit(const at::Tensor& input, const at::Tensor& weight, const std::vector<int>& int4_weight_values) {
   auto result = tensor_methods::reinterpret_cast_4bit( 
                   bridge::GetXlaTensor(input), bridge::GetXlaTensor(weight), int4_weight_values);
   return bridge::AtenFromXlaTensor(std::move(result));
@@ -1406,7 +1406,7 @@ void InitXlaModuleBindings(py::module m) {
   });
   m.def("_xla_reinterpret_cast_4bit",
         [](const at::Tensor& input, const at::Tensor& weight,
-           const std::vector<int8_t>& int4_weight_values) -> at::Tensor {
+           const std::vector<int>& int4_weight_values) -> at::Tensor {
           at::Tensor result;
           {
             NoGilSection nogil;
