+++ torch_xla/csrc/tensor_methods.cpp
@@ -2367,7 +2367,7 @@ XLATensorPtr dequantize_tensor(const XLATensorPtr& input,
 }
 
 XLATensorPtr reinterpret_cast_4bit(const XLATensorPtr& input, const XLATensorPtr& weight,
-                                   const std::vector<int8_t>& int4_weight_values) {
+                                   const std::vector<int>& int4_weight_values) {
   torch::lazy::NodePtr node = torch::lazy::MakeNode<ReinterpretCast4bit>(
       input->GetIrValue(), weight->GetIrValue(), int4_weight_values);
   return input->CreateFrom(torch::lazy::Value(node));
